Home Automation with the Arduino and the Amazon Echo
This sketch measures the state of a washing machine by monitoring it's current. It then sends the data up to the phant cloud so it can be read and interpreted by an Amazon Echo. Go to the ForceTronics YouTube Channel to view the series entitled: Home Automation with the Arduino and the Amazon Echo